[refactor/207-course-card] Extract CourseCardDetails and centralize card UI labels
Part of #207: presentation block + courseCardLabels lib for future i18n.
